TGC: switch from int to float coordinates and adjust existing layouts

+++ TreeGraphControl/CompactTreeLayout.cs
@@ -12,12 +12,12 @@ namespace TGC
 			maxColumns.Clear();
 		}
 
-		private readonly Dictionary<int, int> maxColumns = new Dictionary<int, int>();
-		protected virtual int maxColumn(int y)
+		private readonly Dictionary<int, float> maxColumns = new Dictionary<int, float>();
+		protected virtual float maxColumn(int y)
 		{
 			if (maxColumns.ContainsKey(y))
 				return maxColumns[y];
-			return -1;
+			return -1f;
 		}
 
 		protected virtual bool isLeafNode(VisualTreeNode node)
@@ -25,13 +25,13 @@ namespace TGC
 			return node.Children.Count(child => child.Node.Visible) == 0;
 		}
 
-		protected virtual void PositionNode(VisualTreeNode node, int depth, int minCol)
+		protected virtual void PositionNode(VisualTreeNode node, int depth, float minCol)
 		{
 			if (!node.Node.Visible)
 				return;
 
 			node.Y = depth;
-			int minX = Math.Max(maxColumn(depth) + 1, minCol);
+			float minX = Math.Max(maxColumn(depth) + 1f, minCol);
 
 			if (isLeafNode(node))
 				node.X = minX;
@@ -39,10 +39,10 @@ namespace TGC
 				var visibleChildren = node.Children.Where(child => child.Node.Visible);
 				int i = 0;
 				foreach (VisualTreeNode child in visibleChildren)
-					PositionNode(child, depth + 1, minX - (visibleChildren.Count() / 2) + i++);
+					PositionNode(child, depth + 1, minX - (visibleChildren.Count() / 2f) + i++);
 
 				node.X = Math.Max(
-					(visibleChildren.Min(child => child.X) + visibleChildren.Max(child => child.X)) / 2,
+					(visibleChildren.Min(child => child.X) + visibleChildren.Max(child => child.X)) / 2f,
 					minX
 				);
 			}
